requirements.txt
streamlit
py2neo
mysql-connector-python
pandas
plotly